Session 2
During this session I designed a mount for the motors, added all of the spacers in assembly, imported and fixed the motors in the assembly. I also created sub-assemblies for both right and left slant mecanum wheels with all of the screws and nuts needed and added everything into the main assembly.