Session 3
During this session I created the plates for the other side of the drivetrain. Also I added the spacers to it and screws together with washers to the whole drivetrain assembly. Also  I connected both sides together, however I still need to connect it in a few more spots as this would not be stable and then I need to add odometers in their locations together with spacers so they do not slide around.