Session 4
During this session I managed to finish the drivetrain with odometers being perfectly in the middle of the drivetrain ensuring maximum accuracy. Also I added both control hubs and a battery which are fully enclosed inside the robot to prevent any contact with the surroundings. Also I added all of the screws to ensure that there are no clearance problems and solved the ones that occurred by cutting out some parts of the drivetrain.